skills/anki/study/pyproject.toml
[project]
name = "anki-study"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "rich>=13.0",
    "psycopg2-binary>=2.9",
    "pytest>=8.0",
    "dbclients",
]

[tool.uv.sources]
dbclients = { path = "/Users/bianders/Brian_Code/dbclients-project" }
